+++ client/package.json
@@ -11,6 +11,7 @@
     "cloudinary": "^2.5.1",
     "cropperjs": "^1.6.2",
     "firebase": "^11.3.1",
+    "framer-motion": "^12.6.5",
     "postcss": "^8.4.49",
     "react": "^18.3.1",
     "react-confetti": "^6.4.0",
@@ -19,8 +20,11 @@
     "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^18.3.1",
     "react-firebase-hooks": "^5.1.1",
+    "react-icons": "^5.5.0",
     "react-router-dom": "^7.0.2",
     "react-scripts": "5.0.1",
+    "react-toastify": "^11.0.5",
+    "react-tooltip": "^5.28.1",
     "tailwindcss": "^3.4.16",
     "tailwindcss-textshadow": "^2.1.3",
     "web-vitals": "^2.1.4",
