+++ client/src/pages/teachers/TeachByParts.css
@@ -42,7 +42,7 @@
 }
 
 .sidebar-item:hover {
-  background: #7B68EE; /* Soft Purple (complementary) */
+  background: #7B68EE; /* Soft Purple */
 }
 
 .sidebar-item.active {
@@ -56,59 +56,92 @@
 
 .main-content {
   flex: 1;
-  padding: 40px;
+  padding: 30px;
   display: flex;
   flex-direction: column;
   gap: 20px;
+  align-items: center;
 }
 
 .main-header {
-  text-align: center;
-}
-
-.main-header h1 {
-  color: #FFFFFF;
-  font-size: 32px;
-  margin: 0;
-}
-
-.main-header p {
-  color: #B3B3B3; /* Light Gray */
-  font-size: 18px;
-  margin: 5px 0 0;
-}
-
-.welcome-section {
-  text-align: center;
+  width: 100%;
   max-width: 600px;
-  margin: 0 auto;
+  text-align: center;
+  padding: 15px;
+  background: linear-gradient(180deg, #1A1A1A, #4D4D4D);
+  border-radius: 12px;
+  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.2);
 }
 
-.welcome-section h2 {
-  color: #E6E6E6;
-  font-size: 28px;
-  margin-bottom: 10px;
+.header-title {
+  color: #FFFFFF;
+  font-size: 36px;
+  font-weight: bold;
+  margin: 0;
+  text-shadow: 0 0 8px rgba(51, 153, 255, 0.5); /* Sky Blue glow */
 }
 
-.welcome-section p {
+.header-subtitle {
   color: #B3B3B3;
   font-size: 16px;
-  margin-bottom: 20px;
+  font-style: italic;
+  margin: 5px 0 0;
+  line-height: 1.4;
+}
+
+.welcome-card {
+  background: rgba(128, 128, 128, 0.2); /* Gray with glassmorphism */
+  backdrop-filter: blur(10px);
+  border-radius: 12px;
+  padding: 20px;
+  max-width: 500px;
+  width: 100%;
+  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.3);
+  display: flex;
+  flex-direction: column;
+  gap: 15px;
+}
+
+.welcome-title {
+  color: #E6E6E6;
+  font-size: 28px;
+  margin: 0;
+  position: relative;
+}
+
+.welcome-title::after {
+  content: '';
+  position: absolute;
+  bottom: -5px;
+  left: 0;
+  width: 50px;
+  height: 2px;
+  background: #99CCFF; /* Pale Blue */
+}
+
+.welcome-description {
+  color: #B3B3B3;
+  font-size: 16px;
+  line-height: 1.5;
+  margin: 0;
 }
 
 .start-button {
   background: #3399FF; /* Sky Blue */
   border: none;
   border-radius: 8px;
-  padding: 12px 24px;
+  padding: 10px 20px;
   color: #FFFFFF;
   font-size: 16px;
   cursor: pointer;
-  transition: background 0.3s ease;
+  align-self: flex-end;
+  box-shadow: 0 0 10px rgba(51, 153, 255, 0.5);
+  transition: background 0.3s ease, box-shadow 0.3s ease;
 }
 
 .start-button:hover {
   background: #4DA8FF;
+  box-shadow: 0 0 15px rgba(51, 153, 255, 0.7);
 }
 
 .upload-container {
@@ -116,7 +149,7 @@
   padding: 20px;
   border-radius: 12px;
   max-width: 800px;
-  margin: 0 auto;
+  margin: 20px auto;
   box-shadow: 0 4px 20px rgba(0, 0, 0, 0.5);
 }
